requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==25.3.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
cachetools==6.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.2.1
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.18.1
feather-format==0.4.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.6.1
gast==0.6.0
gevent==23.7.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.65.5
h11==0.16.0
h5py==3.11.0
idna==3.4
imageio==2.37.0
imbalanced-learn==0.12.3
imblearn==0.0
ipykernel==6.25.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==3.5.0
kiwisolver==1.4.4
libclang==18.1.1
llvmlite==0.41.0
macholib==1.16.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
modulegraph==0.19.6
mpmath==1.3.0
multidict==6.5.0
namex==0.0.8
narwhals==1.43.1
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.0
numpy==1.24.3
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
propcache==0.3.2
protobuf==4.25.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py2app==0.28.8
pyarrow==12.0.1
pycparser==2.21
pydeck==0.9.1
Pygments==2.15.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.0.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.1
random-graph==0.0.0a1
referencing==0.30.2
reportlab==4.2.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
selenium==4.32.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.3
streamlit==1.46.0
sympy==1.12
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
torch==2.4.0
torch-geometric==2.6.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.20.1
zope.event==5.0
zope.interface==6.0
